Allow use of Deferred-Sum model in Lzp3 compressor.
A quick test showed that the deferred sum model is actually 43% slower
than the move-to-front model we were previously using.  But decoding
is 38% faster.  Hmph.  Maybe the results would be closer if we used
the deferred-sum model only in contexts not likely to be sparse (ie,
where we expect to see almost every possible symbol at least once).